Refactoring code from merger - Moved old code into Asim's that got merged in Development branch
